Remove `allTests` property and tests
