+++ Tests/OctoKitTests/ConfigurationTests.swift
@@ -3,17 +3,6 @@ import Foundation
 import OctoKit
 
 class ConfigurationTests: XCTestCase {
-    static var allTests = [
-        ("testTokenConfiguration", testTokenConfiguration),
-        ("testTokenConfigurationWithPreviewHeaders", testTokenConfigurationWithPreviewHeaders),
-        ("testEnterpriseTokenConfiguration", testEnterpriseTokenConfiguration),
-        ("testOAuthConfiguration", testOAuthConfiguration),
-        ("testOAuthTokenConfiguration", testOAuthTokenConfiguration),
-        ("testAccessTokenFromResponse", testAccessTokenFromResponse),
-        ("testHandleOpenURL", testHandleOpenURL),
-        ("testLinuxTestSuiteIncludesAllTests", testLinuxTestSuiteIncludesAllTests)
-    ]
-    
     func testTokenConfiguration() {
         let subject = TokenConfiguration("12345")
         XCTAssertEqual(subject.accessToken, "12345".data(using: .utf8)?.base64EncodedString())
@@ -67,18 +56,5 @@ class ConfigurationTests: XCTestCase {
         }
         XCTAssertEqual(token, "017ec60f4a182".data(using: .utf8)?.base64EncodedString())
         XCTAssertTrue(session.wasCalled)
-    }
-    
-    func testLinuxTestSuiteIncludesAllTests() {
-        #if os(macOS) || os(iOS) || os(tvOS) || os(watchOS)
-        let thisClass = type(of: self)
-        let linuxCount = thisClass.allTests.count
-        #if os(iOS)
-        let darwinCount = thisClass.defaultTestSuite.tests.count
-        #else
-        let darwinCount = thisClass.defaultTestSuite.tests.count
-        #endif
-        XCTAssertEqual(linuxCount, darwinCount, "\(darwinCount - linuxCount) tests are missing from allTests")
-        #endif
     }
 }
