+++ Tests/OctoKitTests/FollowTests.swift
@@ -2,18 +2,6 @@ import XCTest
 import OctoKit
 
 class FollowTests: XCTestCase {
-    static var allTests = [
-        ("testGetMyFollowers", testGetMyFollowers),
-        ("testFailToGetMyFollowers", testFailToGetMyFollowers),
-        ("testGetUsersFollowers", testGetUsersFollowers),
-        ("testFailToGetUsersFollowers", testFailToGetUsersFollowers),
-        ("testGetMyFollowing", testGetMyFollowing),
-        ("testFailToGetMyFollowing", testFailToGetMyFollowing),
-        ("testGetUsersFollowing", testGetUsersFollowing),
-        ("testFailToGetUsersFollowing", testFailToGetUsersFollowing),
-        ("testLinuxTestSuiteIncludesAllTests", testLinuxTestSuiteIncludesAllTests)
-    ]
-    
     func testGetMyFollowers() {
         let config = TokenConfiguration("user:12345")
         let headers = Helper.makeAuthHeader(username: "user", password: "12345")
@@ -140,18 +128,5 @@ class FollowTests: XCTestCase {
         }
         XCTAssertNotNil(task)
         XCTAssertTrue(session.wasCalled)
-    }
-    
-    func testLinuxTestSuiteIncludesAllTests() {
-        #if os(macOS) || os(iOS) || os(tvOS) || os(watchOS)
-        let thisClass = type(of: self)
-        let linuxCount = thisClass.allTests.count
-        #if os(iOS)
-        let darwinCount = thisClass.defaultTestSuite.tests.count
-        #else
-        let darwinCount = thisClass.defaultTestSuite.tests.count
-        #endif
-        XCTAssertEqual(linuxCount, darwinCount, "\(darwinCount - linuxCount) tests are missing from allTests")
-        #endif
     }
 }
