+++ Tests/OctoKitTests/PublicKeyTests.swift
@@ -3,11 +3,6 @@ import RequestKit
 import OctoKit
 
 class PublicKeyTests: XCTestCase {
-    static var allTests = [
-        ("testPostPublicKey", testPostPublicKey),
-        ("testLinuxTestSuiteIncludesAllTests", testLinuxTestSuiteIncludesAllTests)
-    ]
-    
     // MARK: Actual Request tests
 
     func testPostPublicKey() {
@@ -25,18 +20,5 @@ class PublicKeyTests: XCTestCase {
         }
         XCTAssertNotNil(task)
         XCTAssertTrue(session.wasCalled)
-    }
-    
-    func testLinuxTestSuiteIncludesAllTests() {
-        #if os(macOS) || os(iOS) || os(tvOS) || os(watchOS)
-        let thisClass = type(of: self)
-        let linuxCount = thisClass.allTests.count
-        #if os(iOS)
-        let darwinCount = thisClass.defaultTestSuite.tests.count
-        #else
-        let darwinCount = thisClass.defaultTestSuite.tests.count
-        #endif
-        XCTAssertEqual(linuxCount, darwinCount, "\(darwinCount - linuxCount) tests are missing from allTests")
-        #endif
     }
 }
