+++ Tests/OctoKitTests/ReactionsTests.swift
@@ -2,11 +2,6 @@ import XCTest
 import OctoKit
 
 class ReactionsTests: XCTestCase {
-    static var allTests = [
-        ("testFullResponse", testFullResponse),
-        ("testLinuxTestSuiteIncludesAllTests", testLinuxTestSuiteIncludesAllTests)
-    ]
-
     // MARK: Model Tests
 
     func testFullResponse() {
@@ -21,18 +16,5 @@ class ReactionsTests: XCTestCase {
         XCTAssertEqual(subject.rocket, 6)
         XCTAssertEqual(subject.eyes, 2)
         XCTAssertEqual(subject.url, "https://api.github.com/repos/octocat/Hello-World/issues/1347/reactions")
-    }
-
-    func testLinuxTestSuiteIncludesAllTests() {
-        #if os(macOS) || os(iOS) || os(tvOS) || os(watchOS)
-        let thisClass = type(of: self)
-        let linuxCount = thisClass.allTests.count
-        #if os(iOS)
-        let darwinCount = thisClass.defaultTestSuite.tests.count
-        #else
-        let darwinCount = thisClass.defaultTestSuite.tests.count
-        #endif
-        XCTAssertEqual(linuxCount, darwinCount, "\(darwinCount - linuxCount) tests are missing from allTests")
-        #endif
     }
 }
