+++ Tests/OctoKitTests/RepositoryTests.swift
@@ -2,19 +2,6 @@ import XCTest
 import OctoKit
 
 class RepositoryTests: XCTestCase {
-    static var allTests = [
-        ("testGetRepositories", testGetRepositories),
-        ("testGetRepositoriesEnterprise", testGetRepositoriesEnterprise),
-        ("testGetAuthenticatedRepositories", testGetAuthenticatedRepositories),
-        ("testGetAuthenticatedRepositoriesEnterprise", testGetAuthenticatedRepositoriesEnterprise),
-        ("testFailToGetRepositories", testFailToGetRepositories),
-        ("testGetRepository", testGetRepository),
-        ("testGetRepositoryEnterprise", testGetRepositoryEnterprise),
-        ("testFailToGetRepository", testFailToGetRepository),
-        ("testUserParsingFullRepository", testUserParsingFullRepository),
-        ("testLinuxTestSuiteIncludesAllTests", testLinuxTestSuiteIncludesAllTests)
-    ]
-    
     // MARK: Actual Request tests
 
     func testGetRepositories() {
@@ -165,18 +152,5 @@ class RepositoryTests: XCTestCase {
         XCTAssertEqual(subject.sshURL, "git@github.com:mietzmithut/Test.git")
         XCTAssertEqual(subject.cloneURL, "https://github.com/mietzmithut/Test.git")
         XCTAssertEqual(subject.size, 132)
-    }
-    
-    func testLinuxTestSuiteIncludesAllTests() {
-        #if os(macOS) || os(iOS) || os(tvOS) || os(watchOS)
-        let thisClass = type(of: self)
-        let linuxCount = thisClass.allTests.count
-        #if os(iOS)
-        let darwinCount = thisClass.defaultTestSuite.tests.count
-        #else
-        let darwinCount = thisClass.defaultTestSuite.tests.count
-        #endif
-        XCTAssertEqual(linuxCount, darwinCount, "\(darwinCount - linuxCount) tests are missing from allTests")
-        #endif
     }
 }
