+++ Tests/OctoKitTests/StarsTests.swift
@@ -2,14 +2,6 @@ import XCTest
 import OctoKit
 
 class StarsTests: XCTestCase {
-    static var allTests = [
-        ("testGetStarredRepositories", testGetStarredRepositories),
-        ("testFailToGetStarredRepositories", testFailToGetStarredRepositories),
-        ("testGetUsersStarredRepositories", testGetUsersStarredRepositories),
-        ("testFailToGetUsersStarredRepositories", testFailToGetUsersStarredRepositories),
-        ("testLinuxTestSuiteIncludesAllTests", testLinuxTestSuiteIncludesAllTests)
-    ]
-    
     // MARK: Actual Request tests
 
     func testGetStarredRepositories() {
@@ -74,18 +66,5 @@ class StarsTests: XCTestCase {
         }
         XCTAssertNotNil(task)
         XCTAssertTrue(session.wasCalled)
-    }
-    
-    func testLinuxTestSuiteIncludesAllTests() {
-        #if os(macOS) || os(iOS) || os(tvOS) || os(watchOS)
-        let thisClass = type(of: self)
-        let linuxCount = thisClass.allTests.count
-        #if os(iOS)
-        let darwinCount = thisClass.defaultTestSuite.tests.count
-        #else
-        let darwinCount = thisClass.defaultTestSuite.tests.count
-        #endif
-        XCTAssertEqual(linuxCount, darwinCount, "\(darwinCount - linuxCount) tests are missing from allTests")
-        #endif
     }
 }
