+++ Tests/OctoKitTests/UserTests.swift
@@ -2,16 +2,6 @@ import XCTest
 import OctoKit
 
 class UserTests: XCTestCase {
-    static var allTests = [
-        ("testGetUser", testGetUser),
-        ("testFailingToGetUser", testFailingToGetUser),
-        ("testGettingAuthenticatedUser", testGettingAuthenticatedUser),
-        ("testFailToGetAuthenticatedUser", testFailToGetAuthenticatedUser),
-        ("testUserParsingFullUser", testUserParsingFullUser),
-        ("testUserParsingMinimalUser", testUserParsingMinimalUser),
-        ("testLinuxTestSuiteIncludesAllTests", testLinuxTestSuiteIncludesAllTests)
-    ]
-    
     // MARK: Actual Request tests
 
     func testGetUser() {
@@ -137,18 +127,5 @@ class UserTests: XCTestCase {
         XCTAssertEqual(subject.numberOfPublicRepos, 7)
         XCTAssertEqual(subject.numberOfPublicGists, 0)
         XCTAssertNil(subject.numberOfPrivateRepos)
-    }
-    
-    func testLinuxTestSuiteIncludesAllTests() {
-        #if os(macOS) || os(iOS) || os(tvOS) || os(watchOS)
-        let thisClass = type(of: self)
-        let linuxCount = thisClass.allTests.count
-        #if os(iOS)
-        let darwinCount = thisClass.defaultTestSuite.tests.count
-        #else
-        let darwinCount = thisClass.defaultTestSuite.tests.count
-        #endif
-        XCTAssertEqual(linuxCount, darwinCount, "\(darwinCount - linuxCount) tests are missing from allTests")
-        #endif
     }
 }
